+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import wooteco.subway.controller.converter.StationConverter;
-import wooteco.subway.dto.controller.request.StationRequest;
-import wooteco.subway.dto.controller.response.StationResponse;
-import wooteco.subway.dto.service.StationDto;
+import wooteco.subway.controller.dto.request.StationRequest;
+import wooteco.subway.controller.dto.response.StationResponse;
 import wooteco.subway.service.StationService;
+import wooteco.subway.service.dto.StationDto;
 
 @RestController
 @RequestMapping("/stations")
