+++ src/main/java/wooteco/subway/controller/converter/PathConverter.java
@@ -3,10 +3,10 @@ package wooteco.subway.controller.converter;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import wooteco.subway.dto.controller.response.PathResponse;
-import wooteco.subway.dto.controller.response.StationResponse;
-import wooteco.subway.dto.service.request.PathServiceRequest;
-import wooteco.subway.dto.service.response.PathServiceResponse;
+import wooteco.subway.controller.dto.response.PathResponse;
+import wooteco.subway.controller.dto.response.StationResponse;
+import wooteco.subway.service.dto.request.PathServiceRequest;
+import wooteco.subway.service.dto.response.PathServiceResponse;
 
 public class PathConverter {
     public static PathServiceRequest toInfo(Long source, Long target, int age) {
