+++ src/main/java/wooteco/subway/dao/JdbcLineDao.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 @Repository
 public class JdbcLineDao implements LineDao {
