+++ src/main/java/wooteco/subway/dao/JdbcSectionDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import wooteco.subway.domain.section.Section;
-import wooteco.subway.dto.SectionEntity;
+import wooteco.subway.entity.SectionEntity;
 
 @Repository
 public class JdbcSectionDao implements SectionDao {
