+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -3,7 +3,7 @@ package wooteco.subway.dao;
 import java.util.List;
 
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 public interface LineDao {
     LineEntity save(Line line);
