+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -3,7 +3,7 @@ package wooteco.subway.dao;
 import java.util.List;
 
 import wooteco.subway.domain.section.Section;
-import wooteco.subway.dto.SectionEntity;
+import wooteco.subway.entity.SectionEntity;
 
 public interface SectionDao {
     void save(Long lineId, Section section);
