+++ src/main/java/wooteco/subway/service/dto/StationDto.java
@@ -1,0 +1,23 @@
+package wooteco.subway.service.dto;
+
+public class StationDto {
+    private Long id;
+    private String name;
+
+    public StationDto(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public StationDto(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
