+++ src/main/java/wooteco/subway/service/dto/request/PathServiceRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.service.request;
+package wooteco.subway.service.dto.request;
 
 public class PathServiceRequest {
     private final long source;
