+++ src/test/java/wooteco/subway/acceptance/PathAcceptanceTest.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import wooteco.subway.dto.controller.response.PathResponse;
+import wooteco.subway.controller.dto.response.PathResponse;
 
 public class PathAcceptanceTest extends AcceptanceTest {
