+++ src/test/java/wooteco/subway/service/FakeLineDao.java
@@ -10,7 +10,7 @@ import org.springframework.util.ReflectionUtils;
 
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.dto.LineEntity;
+import wooteco.subway.entity.LineEntity;
 
 public class FakeLineDao implements LineDao {
     private Long seq = 0L;
